l10n: git.pot: proposed updates for v2.15.0 (+31)
The 1st round of Git l10n often starts very late duing the release
cycle of Git, usually after the release of rc0. This leads to a
gaint update of "po/git.pot" and lots of works for translators with
a tight deadline (about one week).

This "pu" branch holds proposed updates of the file "po/git.pot" for
Git next release, and will be updated and rewound without notification
when there are remarkable updates for it.

Translators could work on this "pu" branch before the l10n coordinator
starts the 1st round of l10n for this release cycle of Git, then
cherry-pick or rebase their work(s) onto the master branch, and send
pull request to the l10n coordinator later.

Please note:

 1. Because this branch will be rewound occasionally, translators
    SHOULD NOT merge, but SHOULD rebase their work(s) to the rewound
    "pu" branch instead.

        $ git fetch git://github.com/git-l10n/git-po.git pu
        $ git rebase --onto FETCH_HEAD HEAD~1

 2. Update your translated message file "po/XX.po".

        $ msgmerge --add-location --backup=off -U "po/XX.po" \
          po/git.pot

 3. Then start to translate "po/XX.po" and commit.

 4. DO NOT send pull request to the l10n coordinator for
    translations on this "pu" branch.

 5. When the l10n coordinator starts the 1st round l10n for the
    current Git release cycle, translators should rebase their works
    onto the master branch, review the translations and then send pull
    request.

        $ git fetch git://github.com/git-l10n/git-po.git master
        $ git rebase --onto FETCH_HEAD HEAD~1

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>